feedForward working, but not correctly

+++ inc/input_parse.h
@@ -12,6 +12,6 @@
 int32_t read_int(std::ifstream &file);
 
 // Function to load IDX3 file and parse images
-std::vector<std::vector<int>> load_IDX3(const std::string &filename);
+std::vector<std::vector<double>> load_IDX3(const std::string &filename);
 
 std::vector<int> load_IDX1_to_array(const std::string &filename, size_t num_labels);
